Move MaybePromptedCString, as it is only used by the piv feature.
And, pull in a newer version of `multitest.py` so we can test this
change. (The old version had a bug which didn't properly support
multi-crate workspaces.)
