+++ yubirs/src/piv/pkey.rs
@@ -15,7 +15,6 @@
 use crate::error::*;
 use crate::piv::id::Algorithm;
 use crate::piv::util::*;
-use crate::util::MaybePromptedCString;
 use failure::format_err;
 use lazy_static::lazy_static;
 use openssl;
